Add bunch of project work.
Starting from here, project work will be more organized.

Work from here will be focused on:

- Creating different networks to learn different representations of
feature space.
- Using/creating different regularization functions.
- Visualizing what different neurons look for.

In particular, I'm going to aim for creating a convolutional network. A
lot of the neat weird things that I liked about DeepDream are, I
realize, convolutional in nature. Different things appeared in different
places, meaning that whatever the network saw, it had to be immune to
translations. Thus, to really get what I want, I have to create a
network that's can see the same things in different areas, however that
happens.

Changes to be committed:
      modified:   own-keras-autoencoder.py